requirements.txt
mysql
lxml
python-dotenv
openai
flask
pymongo
mysql-connector-python
pinecone
